Исправление пространства имен интерфейсов и сервисов
