+++ Services/WebStore.Interfaces/Services/IEmployeesData.cs
@@ -1,6 +1,6 @@
 ﻿using WebStore.Domain.Entities;
 
-namespace WebStore.Services.Interfaces
+namespace WebStore.Interfaces.Services
 {
     public interface IEmployeesData
     {
