+++ Services/WebStore.Interfaces/Services/IProductData.cs
@@ -1,7 +1,7 @@
 ﻿using WebStore.Domain;
 using WebStore.Domain.Entities;
 
-namespace WebStore.Services.Interfaces
+namespace WebStore.Interfaces.Services
 {
     public interface IProductData
     {
