+++ Services/WebStore.Services/Mapping/ProductMapper.cs
@@ -1,7 +1,7 @@
 ﻿using WebStore.Domain.Entities;
 using WebStore.ViewModels;
 
-namespace WebStore.Infrastructure.Mapping
+namespace WebStore.Services.Mapping
 {
     public static class ProductMapper
     {
