+++ Services/WebStore.Services/Services/InMemory/InMemoryProductData.cs
@@ -1,7 +1,7 @@
-﻿using WebStore.Data;
-using WebStore.Domain;
+﻿using WebStore.Domain;
 using WebStore.Domain.Entities;
-using WebStore.Services.Interfaces;
+using WebStore.Interfaces.Services;
+using WebStore.Services.Data;
 
 namespace WebStore.Services
 {
