+++ UI/WebStore/Areas/Admin/Controllers/ProductsController.cs
@@ -2,8 +2,8 @@
 using Microsoft.AspNetCore.Mvc;
 using WebStore.Areas.Admin.ViewModels;
 using WebStore.Domain.Entities.Identity;
-using WebStore.Infrastructure.Mapping;
-using WebStore.Services.Interfaces;
+using WebStore.Interfaces.Services;
+using WebStore.Services.Mapping;
 
 namespace WebStore.Areas.Admin.Controllers
 {
