+++ UI/WebStore/Program.cs
@@ -4,10 +4,10 @@ using WebStore.DAL.Context;
 using WebStore.Domain.Entities.Identity;
 using WebStore.Infrastructure.Conventions;
 using WebStore.Infrastructure.Middleware;
-using WebStore.Services;
-using WebStore.Services.InCookies;
-using WebStore.Services.InSQL;
-using WebStore.Services.Interfaces;
+using WebStore.Interfaces.Services;
+using WebStore.Services.Services;
+using WebStore.Services.Services.InCookies;
+using WebStore.Services.Services.InSQL;
 
 var builder = WebApplication.CreateBuilder(args); //Создание построителя приложения
